+++ lib/dopstick.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "json"
 require "date"
 require "pathname"
 require "thor"
@@ -11,5 +12,7 @@ module Dopstick
   require_relative "dopstick/generator/options"
   require_relative "dopstick/generator/gem/options"
   require_relative "dopstick/generator/gem/generator"
+  require_relative "dopstick/generator/npm/options"
+  require_relative "dopstick/generator/npm/generator"
   require_relative "dopstick/cli"
 end